src/requirements.txt
strsim
imdbpy